Map some input filenames to have a PDF extension
We discovered that from Google Draw, you get better clarity on the image
exports when you export to PDF. We still have the filenames referencing
the JPG images in the source, though (which is still important for HTML
output), so we map the appropriate names to have a PDF extension at the
XSLT processing level. This includes some XSLT gymnastics to create
appropriate node-sets for mapping (while also maintaining the ancestor
trail of any xml:base attributes). This also includes a hard-coded
whitelist of filenames not to modify.